Moved iterator header.
