+++ encoding_lib/src/adaptive_a_dictionary.cpp
@@ -1,6 +1,6 @@
 #include <ael/dictionary/adaptive_a_dictionary.hpp>
 
-#include <ael/dictionary/integer_random_access_iterator.hpp>
+#include "integer_random_access_iterator.hpp"
 #include <boost/range/iterator_range.hpp>
 
 namespace ael::dict {
